requirements.txt
Django
djangorestframework
celery[redis]
dj-database-url
python-decouple
django-celery-beat
django-extensions
gunicorn
psycopg2-binary
sentry-sdk
requests
python-magic
openai-whisper
openai
